Changed the jumpscare timeline
READ THIS IT'S IMPORTANT:

the jumpscare timeline isn't done yet. there are some bugs with the placement of the player and the monster

after you are done with that, make sure to fix the double door bug. hint: after you collide with a door, turn of all door collision components for 2 seconds to make sure that we dont double collide