Added car for night 1 scene
Here's the plan for night 1:

You spawn in a car and you are driving on a road. Suddenly, theres fire on your hood and you crash into the forest. You see a trail and you have to follow the trail or you die. Next, if you don't get to the house in the distance in time, you die. Next, if you hear rustling in the woods, you stop and play an audio, and then continue.