Did lots of work on the zooming in and out feature
Haven't actually added any functionality to the task buttons, but will do that tomorrow or today

Most likely tomorrow

Day two is getting very good so far. had some doubts at the very start of day 2 because of all the bugs and confusions, but it looks good now. smooth sailings with the development